chore: move validation to its own file
